Implement topological sort of locked dependencies
Dependencies are sorted topologically before writing them into the lock
file. When the lock file is loaded the order is preserved by using
`OrderedTable` data structure and they will be installed in a such way,
that if some dependency must be build before installing, its
dependencies will be already installed.

If a cycle in the dependency graph is detected then a warning will be
printed, but the algorithm will continue, because it is not expected
the most of the dependencies to be binary and to require build. Despite
the fact that the cyclic dependencies are invalid, there are many cases
that they will actually work and it is better to continue with a warning
which should be fixed by package developers, instead of breaking the
build prematurely.

Unit tests for the topological sort are added and the "Module tests"
suite in `tester.nim` file is re-factored to use template for minimizing
the code duplication.

Related to nim-lang/nimble#127