Split tests in separate files by suite
The tests from the `tester.nim` module are split into separate files for
each of the test suites in order to achieve better maintainability of
the code and faster compilation when only one of the suites is modified.
All new modules are imported from the `tester.nim` module and the common
code between them is moved in a new `testscommon.nim` module.

Related to nim-lang/Nimble#127